Avoid calling getThemedContext() to inflate our custom ActionBar view
Although the ActionBar is available from API level 11 onwards, the
getThemedContext() call is not available until API 14, causing the
crash reported in issue #30. This is a fatal build error with newer
SDKs (which now have the method marked as level 14 correctly).

Fortunately, the ActionBar can also inflate a view from a resource
id so we don't have to ask it for the inflator. We can get the inflated
view back to manipulate it, avoiding the use of getThemedContext()
entirely.

With a bit of refactoring, the code ends up simpler despite having to
do things this way.